Made meta balls
The demo shows two meta balls moving, joining, and seperating, but it is terribly inefficient. The problem could be solved by using points and textures rather than geometries, but that wont work for proper marching squares, and it wont help to actually increase the efficiency much. The main problem is that add needs to search through the entire list to not add duplicates, and remove needs to remake the list of sectors filled whenever it removes one, or does it?