Began refactoring for making game
Because this is a game, I want to do some refactoring so I am not messign around with matrices and stuff in the main runner of the program. I want to abstract it away in a way that makes sense, since the Mesh class by itself does not carry that much information.